Update code.js
I was stumped on how to find all the permutations, so I went to look at the work of the people who had done this assignment before me. And the only one that I found was. https://github.com/COSC3020/detecting-isomorphism-cadenmcfate 
For his solution, he uses a very specific type of adjacency list, and it will only work with this specific type of adjacency list. ( I tried very hard to adapt it into a normal adjacency list.) So, I took his special adjacent list and my general understanding of what his program did, and tried to make my own.